// src/app/models/application.model.ts
 export class Applicant {
     id: number;
     userId: number;
     resumeId: number;
     salary: number;
     firstName: string;
     lastName: string;
     gender: string;
     dateApplied: Date;
     job: string;
     address: string;
     city: string;
     state: string;
     zip: number;
 }